raise limit of varint to u16, use associated const

+++ src/multihash.rs
@@ -249,7 +249,7 @@ where
     let code = read_u64(&mut r)?;
     let size = read_u64(&mut r)?;
 
-    if size > S as u64 || size > u8::MAX as u64 {
+    if size > S as u64 || size > u16::MAX as u64 {
         return Err(Error::InvalidSize(size));
     }
